moved auth variables into auth.js
