--- src/app/Auth/auth0-variables.js
@@ -1,6 +1,0 @@
-/* eslint-disable-next-line import/prefer-default-export */
-export const AUTH_CONFIG = {
-    domain: 'revsearch.eu.auth0.com',
-    clientId: 'i9DlcDL1TASkNDhhFaZOXIsx996ySFZ6',
-    callbackUrl: 'http://localhost:8081/callback',
-};
